requirements.txt
# Required Python packages for the Air Pollution Monitoring System

# Core data handling
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP requests for API calls
requests>=2.31.0

# Google Earth Engine (for satellite data)
earthengine-api>=0.1.370

# Data visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment variables (for API keys)
python-dotenv>=1.0.0
